Add workaround to allow users on desktop safari to download files even if pop up blocking is enabled.
NB: Safari on mobile does not support the download attribute and so you will need to have the "block pop ups" setting disabled in Settings -> Safari if attempting to download files on mobile devices. This is because safari on mobile does not have access to the file storage and thus cannot download things directly to the device.